+++ gazebo/util/LogPlay.cc
@@ -300,6 +300,20 @@ bool LogPlay::Step(std::string &_data)
 }
 
 /////////////////////////////////////////////////
+bool LogPlay::Rewind()
+{
+  this->currentChunk.clear();
+  this->logCurrXml = this->logStartXml->FirstChildElement("chunk");
+  if (!logCurrXml)
+  {
+    gzerr << "Unable to jump to the beginning of the log file\n";
+    return false;
+  }
+
+  return true;
+}
+
+/////////////////////////////////////////////////
 bool LogPlay::GetChunk(unsigned int _index, std::string &_data)
 {
   unsigned int count = 0;